Remove circuit breaker docs, functions, and config
The last remaining circuit breaker was used in the Postgres notifier,
which no longer manages its own connection. Instead, the persistent
connection manages auto reconnection by itself and between the
`Connection.connected?` and `Senator.leader?` callbacks, plugins or
other connection sensitve code can tailor behavior.
